requirements.txt
dj_database_url==0.5.0
Django==3.1.14
django_heroku==0.3.1
ipython==8.3.0
numpy==1.20.3
pandas==1.3.4
Pillow==9.1.0
psycopg2_binary==2.9.1
SQLAlchemy==1.4.22
keras==2.8.0
